extract old password component
